+++ src/components/aside/modal/Modal.jsx
@@ -1,16 +1,16 @@
-import styles from "/src/styles/Aside.module.css";
+import styles from "/src/styles/Modal.module.css";
 
 export default function Modal({ title, children, onClose, onSubmit, showCloseButton = true }) {
     return (
         <div className={`${styles["modal"]} ${styles["modal--open"]}`}>
-            <div className={styles["modal-backdrop"]} onClick={onClose}></div>
-            <div className={styles["modal-container"]}>
+            <div className={`${styles["modal-backdrop"]}`} onClick={onClose}></div>
+            <div className={`${styles["modal-container"]}`}>
                 {title && <h2 className={`${styles["modal-title"]} text-title`}>{title}</h2>}
                 <form onSubmit={onSubmit || ((e) => e.preventDefault())}>
                     {children}
                     {showCloseButton && (
-                        <div className={styles["button-container"]}>
-                            <button type={onSubmit ? "submit" : "button"} onClick={!onSubmit ? onClose : undefined} className={`${styles.button} ${styles["button--primary"]} text-caption`}>
+                        <div className={`${styles["button-container"]}`}>
+                            <button type={onSubmit ? "submit" : "button"} onClick={!onSubmit ? onClose : undefined} className={`${styles["button"]} ${styles["button--primary"]} text-caption`}>
                                 {onSubmit ? "추가하기" : "닫기"}
                             </button>
                         </div>
