+++ src/components/main/RestaurantList.jsx
@@ -1,12 +1,12 @@
-import styles from "/src/styles/Main.module.css"
 import RestaurantValue from "./RestaurantValue";
+import styles from "/src/styles/RestaurantList.module.css"
 
   const filteredRestaurants = selectedCategory === '전체' ? restaurants : restaurants.filter(restaurant => restaurant.alt === selectedCategory);
 export default function RestaurantList({ restaurants, selectedCategory, setIsModalOpen, setRestaurantValue }) {
 
   return (
-    <section className={styles["restaurant-list-container"]}>
-      <ul className={styles["restaurant-list"]}>
+    <section className={`${styles["restaurant-list-container"]}`}>
+      <ul className={`${styles["restaurant-list"]}`}>
         {filteredRestaurants.map((restaurant) => (
           <RestaurantValue
             key={restaurant.id}
@@ -20,5 +20,5 @@ export default function RestaurantList({ restaurants, selectedCategory, setIsMod
         ))}
       </ul>
     </section>
-   );
+  );
 }
