--- src/styles/Head.module.css
@@ -1,33 +1,0 @@
-/* Header ********************************************/
-.gnb {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    height: 64px;
-  
-    padding: 0 16px;
-  
-    background-color: var(--primary-color);
-  }
-  
-  .gnb__title {
-    color: #fcfcfd;
-  }
-  
-  .gnb__button {
-    height: 40px;
-  
-    border: none;
-    border-radius: 8px;
-    background: transparent;
-  
-    font-size: 24px;
-    cursor: pointer;
-  }
-  
-  .gnb__button img {
-    display: block;
-    width: 40px;
-    height: 40px;
-    object-fit: contain;
-  }
