+++ src/styles/Modal.module.css
@@ -1,0 +1,66 @@
+.modal {
+    display: none;
+}
+
+.modal--open {
+    display: block;
+}
+
+.modal-backdrop {
+    position: fixed;
+    top: 0;
+    right: 0;
+    bottom: 0;
+    left: 0;
+
+    background: rgba(0, 0, 0, 0.35);
+}
+
+.modal-container {
+    position: fixed;
+    bottom: 0;
+    width: 100%;
+
+    padding: 32px 16px;
+
+    border-radius: 8px 8px 0px 0px;
+    background: var(--grey-100);
+}
+
+.modal-title {
+    margin-bottom: 36px;
+}
+
+.button-container {
+    display: flex;
+}
+
+.button {
+    width: 100%;
+    height: 44px;
+
+    margin-right: 16px;
+
+    border: none;
+    border-radius: 8px;
+
+    font-weight: 600;
+    cursor: pointer;
+}
+
+.button:last-child {
+    margin-right: 0;
+}
+
+.button--secondary {
+    border: 1px solid var(--grey-300);
+    background: transparent;
+
+    color: var(--grey-300);
+}
+
+.button--primary {
+    background: var(--primary-color);
+
+    color: var(--grey-100);
+}
