+++ src/styles/RestaurantDetailModal.module.css
@@ -1,0 +1,4 @@
+/* 음식점 정보 모달 */
+.restaurant-info {
+    margin-bottom: 24px;
+}
