+++ src/styles/RestaurantFilter.module.css
@@ -1,0 +1,23 @@
+/* 카테고리/정렬 필터 */
+.restaurant-filter-container {
+    display: flex;
+    justify-content: space-between;
+
+    padding: 0 16px;
+    margin-top: 24px;
+}
+
+.restaurant-filter-container select {
+    height: 44px;
+    min-width: 125px;
+
+    border: 1px solid #d0d5dd;
+    border-radius: 8px;
+    background: transparent;
+
+    font-size: 16px;
+}
+
+.restaurant-filter {
+    padding: 8px;
+}
