+++ src/styles/RestaurantListItem.module.css
@@ -1,0 +1,53 @@
+.restaurant {
+    display: flex;
+    align-items: flex-start;
+
+    padding: 16px 8px;
+
+    border-bottom: 1px solid #e9eaed;
+}
+
+.restaurant__category {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 64px;
+    height: 64px;
+    min-width: 64px;
+    min-height: 64px;
+
+    margin-right: 16px;
+
+    border-radius: 50%;
+    background: var(--lighten-color);
+}
+
+.category-icon {
+    width: 36px;
+    height: 36px;
+}
+
+.restaurant__info {
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
+}
+
+.restaurant__name {
+    margin: 0;
+}
+
+.restaurant__distance {
+    color: var(--primary-color);
+}
+
+.restaurant__description {
+    display: -webkit-box;
+
+    padding-top: 8px;
+
+    overflow: hidden;
+    text-overflow: ellipsis;
+    -webkit-line-clamp: 2;
+    -webkit-box-orient: vertical;
+}
